updated code

+++ cppnew/test.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 using namespace std;
 
-double square_s2(double s2){
+int square_s2(int s2){
     return s2 * s2;
 }
 
-void print_s2(double s2){
+void print_s2(int s2){
     cout << "s2 squared gives the result " << s2 << endl;
 }
